Basic Support to 0.11 - 0.16

+++ src/main/java/cn/nukkit/entity/EntityHumanType.java
@@ -154,7 +154,7 @@ public abstract class EntityHumanType extends EntityCreature implements Inventor
 
             for (Item armor : inventory.getArmorContents()) {
                 armorPoints += armor.getArmorPoints();
-                epf += calculateEnchantmentProtectionFactor(armor, source);
+                epf += (int) calculateEnchantmentProtectionFactor(armor, source);
             }
 
             //float originalDamage = source.getDamage();
